L3: decompose the service bin — 6247 → 1818 lines, ceiling tightened, panic budget to zero
The DD review (§8.3) flagged kirra_verifier_service.rs as the largest
maintainability liability: a ratcheted-but-rising monolith. Two pure
moves take it under the review's <2000 target:

- The 12 binary-internal #[cfg(test)] modules (~4000 lines: posture-gate
  router proofs, fabric feed/command, attestation nonce handlers, DNP3
  audit, console phase A, store offload, federation e2e, industrial
  replay, systemd notify, LockedOut seeding) move to files under
  src/bin/kirra_verifier_service/, one file per module, banners intact.
  store_offload_guard's include_str! source-scan paths re-anchored to
  its new location.
- The posture-feed cluster (verifier→fabric local-asset feed, recalc
  triggers, freshness wiring, posture-event emission) and the systemd
  cluster (sd_notify + Type=notify watchdog) become tracked submodules
  posture_feed.rs / systemd.rs, same #[path] + use-glob pattern as the
  existing handler modules.

Guardrail ratchet TIGHTENED to lock the gain: bin max_lines 6247 → 1818,
bin panic budget 150 → 0 (every panic site in the file was test code —
the production bin is now panic-free by gate). posture_feed.rs (394/0)
and systemd.rs (86/0) join the tracked set.

Verified: 119 bin-internal tests + 2802 workspace tests green, fmt-clean,
guardrails/orphan gates green.
